bot/services/gis_parser_service.py
"""
GIS ЖКХ Parser Service
Parses public data from dom.gosuslugi.ru to extract RSO (Resource Supply Organizations)

WARNING: This is experimental and may break if GIS changes their HTML structure.
For production, consider using official SOAP API (requires accreditation).
"""
import aiohttp
import logging
from typing import List, Dict, Optional
from bs4 import BeautifulSoup
import re


class GISGKHParser:
    """Parser for dom.gosuslugi.ru public pages"""
    
    BASE_URL = "https://dom.gosuslugi.ru"
    
    async def search_house_by_fias(self, fias_id: str) -> Optional[Dict]:
        """
        Search house in GIS ЖКХ by FIAS ID.
        
        Returns dict with:
            - house_id: Internal GIS ID
            - address: Full address
            - uk: Management company info
            - url: Link to house page
        """
        # Note: This is a PLACEHOLDER. Actual implementation requires:
        # 1. Studying the public API (may require auth)
        # 2. Reverse engineering search requests
        # 3. Handling pagination and filters
        
        logging.warning("GIS ЖКХ parser not fully implemented yet")
        return None
    
    async def get_rso_for_house(self, house_id: str) -> List[Dict]:
        """
        Get list of RSO providers for a house.
        
        Args:
            house_id: Internal GIS ЖКХ house ID
        
        Returns:
            List of dicts with keys:
                - name: RSO name
                - inn: Tax ID
                - service_type: Type of service (электроснабжение, водоснабжение, etc.)
                - contract_number: Contract number (if available)
        """
        url = f"{self.BASE_URL}/#!/house-view/{house_id}"
        
        try:
            async with aiohttp.ClientSession() as session:
                async with session.get(url, timeout=15) as resp:
                    if resp.status != 200:
                        logging.error(f"GIS ЖКХ returned {resp.status}")
                        return []
                    
                    html = await resp.text()
                    
                    # Parse HTML (requires inspecting actual page structure)
                    soup = BeautifulSoup(html, 'html.parser')
                    
                    # THIS IS A PLACEHOLDER - actual selectors need to be determined
                    # by inspecting the real HTML structure
                    rso_list = []
                    
                    # Example (NOT REAL):
                    # rso_elements = soup.select('.rso-provider-item')
                    # for elem in rso_elements:
                    #     name = elem.select_one('.provider-name').text
                    #     inn = elem.select_one('.provider-inn').text
                    #     service = elem.select_one('.service-type').text
                    #     rso_list.append({
                    #         'name': name.strip(),
                    #         'inn': inn.strip(),
                    #         'service_type': service.strip()
                    #     })
                    
                    logging.warning("HTML parsing not implemented - requires manual inspection")
                    return rso_list
        
        except Exception as e:
            logging.error(f"GIS ЖКХ parsing failed: {e}")
            return []
    
    async def search_house_by_address(self, address: str) -> Optional[str]:
        """
        Search house by address string, return house_id.
        
        This would require:
        1. Making a search query
        2. Parsing search results
        3. Extracting house_id from result URL
        """
        # PLACEHOLDER
        logging.warning("Address search not implemented")
        return None


# ===== ALTERNATIVE: Use reformagkh.ru (easier!) =====

class ReformaGKHParser:
    """
    Parser for reformagkh.ru - a more parseable alternative to GIS ЖКХ.
    This site has cleaner HTML and may be easier to scrape.
    """
    
    BASE_URL = "https://www.reformagkh.ru"
    
    async def search_house(self, region: str, city: str, street: str, house: str) -> Optional[Dict]:
        """
        Search house on reformagkh.ru
        
        Example URL pattern:
        https://www.reformagkh.ru/search/houses?q=Москва+Ленина+5
        """
        query = f"{city} {street} {house}"
        search_url = f"{self.BASE_URL}/search/houses?q={query}"
        
        try:
            async with aiohttp.ClientSession() as session:
                headers = {
                    'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
                }
                
                async with session.get(search_url, headers=headers, timeout=15) as resp:
                    if resp.status != 200:
                        return None
                    
                    html = await resp.text()
                    soup = BeautifulSoup(html, 'html.parser')
                    
                    # Find first result
                    # NOTE: Actual selectors need to be determined by inspecting the page
                    # This is a TEMPLATE
                    
                    first_result = soup.select_one('.search-result-item')
                    if not first_result:
                        return None
                    
                    # Extract link to house page
                    link = first_result.select_one('a')
                    if not link or not link.get('href'):
                        return None
                    
                    house_url = self.BASE_URL + link['href']
                    
                    # Now fetch house details
                    return await self._parse_house_page(house_url)
        
        except Exception as e:
            logging.error(f"ReformaGKH search failed: {e}")
            return None
    
    async def _parse_house_page(self, url: str) -> Dict:
        """Parse individual house page to extract UK and RSO info"""
        try:
            async with aiohttp.ClientSession() as session:
                headers = {
                    'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
                }
                
                async with session.get(url, headers=headers, timeout=15) as resp:
                    if resp.status != 200:
                        return {}
                    
                    html = await resp.text()
                    soup = BeautifulSoup(html, 'html.parser')
                    
                    # Extract data (TEMPLATE - needs real selectors)
                    result = {
                        'url': url,
                        'address': '',
                        'uk': None,
                        'rso_list': []
                    }
                    
                    # Example (NOT REAL SELECTORS):
                    # address_elem = soup.select_one('.house-address')
                    # if address_elem:
                    #     result['address'] = address_elem.text.strip()
                    
                    # uk_elem = soup.select_one('.uk-name')
                    # if uk_elem:
                    #     result['uk'] = {
                    #         'name': uk_elem.text.strip(),
                    #         'inn': soup.select_one('.uk-inn').text.strip()
                    #     }
                    
                    # rso_elems = soup.select('.rso-provider')
                    # for rso in rso_elems:
                    #     result['rso_list'].append({
                    #         'name': rso.select_one('.rso-name').text,
                    #         'service': rso.select_one('.service-type').text
                    #     })
                    
                    logging.warning("House page parsing not fully implemented")
                    return result
        
        except Exception as e:
            logging.error(f"House page parsing failed: {e}")
            return {}


# ===== Usage Example =====

async def example_usage():
    """Example: How to use parsers"""
    
    # Option 1: GIS ЖКХ (harder, requires FIAS)
    gis_parser = GISGKHParser()
    fias_id = "abc-123-def"  # From DaData
    house_info = await gis_parser.search_house_by_fias(fias_id)
    if house_info:
        house_id = house_info['house_id']
        rso_list = await gis_parser.get_rso_for_house(house_id)
        print("RSO from GIS ЖКХ:", rso_list)
    
    # Option 2: ReformaGKH (easier)
    reforma_parser = ReformaGKHParser()
    house = await reforma_parser.search_house(
        region="Москва",
        city="Москва",
        street="ул. Ленина",
        house="5"
    )
    if house:
        print("UK:", house.get('uk'))
        print("RSO:", house.get('rso_list'))


# ===== IMPORTANT NOTES =====
"""
1. **Legal Compliance:**
   - Check Terms of Service before scraping
   - Respect robots.txt
   - Add delays between requests
   - Identify your bot in User-Agent

2. **Reliability:**
   - HTML structure can change without notice
   - Implement error handling and fallbacks
   - Cache results to reduce load

3. **Alternative Approach:**
   - For production, consider manual data entry via admin panel
   - Or partner with a data provider
   - Or apply for official GIS ЖКХ API access

4. **Next Steps to Complete This:**
   - Manually inspect dom.gosuslugi.ru HTML
   - Find CSS selectors for RSO data
   - Test with real addresses
   - Add retry logic and rate limiting
"""
